moved redis config t seperate file
